+++ src/js/api/state.js
@@ -4,7 +4,6 @@ import {
     data as getData,
     hasOwn,
     hyphenate,
-    includes,
     isArray,
     isFunction,
     isNumeric,
@@ -101,16 +100,11 @@ export default function (UIkit) {
                 }
             }
         }
-
-        if (hasUpdateOfType(this.$options.update, 'scroll')) {
-            registerScrollListener(this._uid, () => this.$emit('scroll'));
-        }
     };
 
     UIkit.prototype._unbindEvents = function () {
         this._events.forEach((unbind) => unbind());
         delete this._events;
-        unregisterScrollListener(this._uid);
     };
 
     UIkit.prototype._initObservers = function () {
@@ -313,33 +307,4 @@ function initPropsObserver(component) {
     });
 
     return observer;
-}
-
-function hasUpdateOfType(updates = [], type) {
-    for (const { events = [] } of updates) {
-        if (includes(events, type)) {
-            return true;
-        }
-    }
-}
-
-const scrollListeners = new Map();
-let unbindScrollListener;
-function registerScrollListener(id, listener) {
-    unbindScrollListener =
-        unbindScrollListener ||
-        on(window, 'scroll', () => scrollListeners.forEach((listener) => listener()), {
-            passive: true,
-            capture: true,
-        });
-
-    scrollListeners.set(id, listener);
-}
-
-function unregisterScrollListener(id) {
-    scrollListeners.delete(id);
-    if (unbindScrollListener && !scrollListeners.size) {
-        unbindScrollListener();
-        unbindScrollListener = null;
-    }
 }